Remove InstanceListActivity
