--- collect_app/src/main/java/org/odk/collect/android/activities/InstanceListActivity.java
@@ -1,37 +1,0 @@
-package org.odk.collect.android.activities;
-
-import org.odk.collect.android.database.instances.DatabaseInstanceColumns;
-
-import static org.odk.collect.android.utilities.ApplicationConstants.SortingOrder.BY_DATE_ASC;
-import static org.odk.collect.android.utilities.ApplicationConstants.SortingOrder.BY_DATE_DESC;
-import static org.odk.collect.android.utilities.ApplicationConstants.SortingOrder.BY_NAME_ASC;
-import static org.odk.collect.android.utilities.ApplicationConstants.SortingOrder.BY_NAME_DESC;
-import static org.odk.collect.android.utilities.ApplicationConstants.SortingOrder.BY_STATUS_ASC;
-import static org.odk.collect.android.utilities.ApplicationConstants.SortingOrder.BY_STATUS_DESC;
-
-abstract class InstanceListActivity extends AppListActivity {
-    protected String getSortingOrder() {
-        String sortingOrder = DatabaseInstanceColumns.DISPLAY_NAME + " COLLATE NOCASE ASC, " + DatabaseInstanceColumns.STATUS + " DESC";
-        switch (getSelectedSortingOrder()) {
-            case BY_NAME_ASC:
-                sortingOrder = DatabaseInstanceColumns.DISPLAY_NAME + " COLLATE NOCASE ASC, " + DatabaseInstanceColumns.STATUS + " DESC";
-                break;
-            case BY_NAME_DESC:
-                sortingOrder = DatabaseInstanceColumns.DISPLAY_NAME + " COLLATE NOCASE DESC, " + DatabaseInstanceColumns.STATUS + " DESC";
-                break;
-            case BY_DATE_ASC:
-                sortingOrder = DatabaseInstanceColumns.LAST_STATUS_CHANGE_DATE + " ASC";
-                break;
-            case BY_DATE_DESC:
-                sortingOrder = DatabaseInstanceColumns.LAST_STATUS_CHANGE_DATE + " DESC";
-                break;
-            case BY_STATUS_ASC:
-                sortingOrder = DatabaseInstanceColumns.STATUS + " ASC, " + DatabaseInstanceColumns.DISPLAY_NAME + " COLLATE NOCASE ASC";
-                break;
-            case BY_STATUS_DESC:
-                sortingOrder = DatabaseInstanceColumns.STATUS + " DESC, " + DatabaseInstanceColumns.DISPLAY_NAME + " COLLATE NOCASE ASC";
-                break;
-        }
-        return sortingOrder;
-    }
-}
